--- src/components/productPage/review/review/TranslationBtn.tsx
@@ -1,17 +1,0 @@
-import { Theme, css } from '@emotion/react';
-
-export const btnStyle = (theme: Theme) => css`
-	${theme.fonts.kor.captionMedium12}
-	color: ${theme.colors.notificationPrimary};
-
-	background-color: transparent;
-	border: none;
-`;
-
-const TranslationBtn = () => (
-	<button type="button" css={btnStyle}>
-		원문표시
-	</button>
-);
-
-export default TranslationBtn;
